Take the Carbon file system out, all but the corner nobody visits
Twenty-five calls to a way of naming files that the system stopped keeping up with in 2012. Three
are left and they are in libezv, which is vendored and meant to stay as it is.

TickCount, eight sites, was the clock everything throttled against: don't redraw the transfer row
more than five times a second, don't update the search status more than three. Replaced by a
monotonic clock counting the same sixtieths of a second, so that every interval, comparison and
stored queue of them reads exactly as it did and only the call changed. Two types had to grow with
it, and one atomic compare-and-swap beside it, which would otherwise have swapped half of a counter
it had been told was whole.

LSGetApplicationForURL, five sites, all asking the same question in the same roundabout way: what
opens an http address, please answer as a file reference, and then the path out of that. The
workspace answers it directly. One of the five had grown a whole Carbon launch specification around
itself to open a file in the browser it found; that is now one message.

Both alias resolutions, in the transcripts folder and in the file manager category, went from four
chained Carbon calls to one NSURL message. Something that is not an alias resolves to nothing and is
handed back unchanged, which is what those chains worked out from a flag on the way past.

And a random string generator stopped seeding itself with the time since boot, which is neither
unpredictable nor uniform, and asks for random numbers rather than sowing them.

Warnings are at 483, from 6033 when this started.

+++ Source/ESFileTransferProgressRow.h
@@ -23,7 +23,7 @@
 	ESFileTransfer			*fileTransfer;
 	id						owner;
 
-	UInt32					lastUpdateTick;
+	uint64_t				lastUpdateTick;
 	unsigned long long		lastBytesSent;
 	unsigned long long		size;
 	NSString				*sizeString;
